Also fix deep imports of child classes
