--- tests/expected_results/enums/v1/schemas/level_one.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-
-from schemas.level_two import LevelTwo
-
-
-class LevelOne(BaseModel):
-    two: LevelTwo
