--- tests/expected_results/enums/v1/schemas/level_three.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class EnumLevelThree(str, Enum):
-    SOMETHING = "something"
-    SOMETHING_ELSE = "something_else"
-    ANOTHER_THING = "another_thing"
