--- tests/expected_results/enums/v2/schemas/level_two.py
@@ -1,11 +1,0 @@
-from typing import Optional
-
-from pydantic import BaseModel
-
-from schemas.level_three import EnumLevelThree
-
-
-class LevelTwo(BaseModel):
-    three: EnumLevelThree
-    some: Optional[EnumLevelThree]
-    others: list[EnumLevelThree]
